Update BabyImmunization and Add WarningAlertMsg

+++ src/Pages/SpecialMsgDisplay.jsx
@@ -13,6 +13,7 @@ import Grid from "@mui/material/Grid";
 import { Link, IconButton } from "@mui/material";
 import PhoneIcon from "@mui/icons-material/Phone";
 import EmailIcon from "@mui/icons-material/Email";
+import SuccessAlert from "../Components/WarningAlertMsg";
 
 const VISIBLE_FIELDS = ["id", "msg_Content", "msg_Date", "msg_Topic"];
 
@@ -55,6 +56,13 @@ export default function SpecialMsgDisplay() {
 
   const handleDelete = async (id) => {
     try {
+      const rowToDelete = customDataset.find((row) => row.id === id);
+      if (rowToDelete && rowToDelete.msg_Priority === "Ug") {
+        // If the message is urgent, do not allow deletion
+        setShowSuccessAlert(true);
+        console.log("Cannot delete urgent message");
+        return;
+      }
       await deleteSpecialMsg(id);
       console.log(id);
       const updatedDataset = customDataset.filter((row) => row.id !== id);
@@ -65,48 +73,49 @@ export default function SpecialMsgDisplay() {
   };
 
   return (
-    <Box sx={{ display: "flex" }}>
-      <div style={{ display: "flex", height: "100vh", overflowX: "hidden" }}>
-        <DisplaySidebar />
-        <div style={{ flex: 1, overflowX: "hidden", paddingLeft: "40px" }}>
-          <div style={{ height: "100vh", width: "100%" }}>
-            <Typography
-              variant="h5"
-              style={{
-                marginBottom: "10px",
-                color: "#2A777C",
-                paddingBottom: "60px",
-                paddingLeft: "10px",
-                paddingTop: "40px",
-              }}
-            >
-              Special Notes
-            </Typography>
-            <div>
-              <Grid container spacing={6} justifyContent="center">
-                <Grid item>
-                  <div className="legend">
-                    <span className="urgency-color"></span>{" "}
-                    <span className="legend-text">Urgent</span>
-                  </div>
+    <>
+      <Box sx={{ display: "flex" }}>
+        <div style={{ display: "flex", height: "100vh", overflowX: "hidden" }}>
+          <DisplaySidebar />
+          <div style={{ flex: 1, overflowX: "hidden", paddingLeft: "40px" }}>
+            <div style={{ height: "100vh", width: "100%" }}>
+              <Typography
+                variant="h5"
+                style={{
+                  marginBottom: "10px",
+                  color: "#2A777C",
+                  paddingBottom: "60px",
+                  paddingLeft: "10px",
+                  paddingTop: "40px",
+                }}
+              >
+                Special Notes
+              </Typography>
+              <div>
+                <Grid container spacing={6} justifyContent="center">
+                  <Grid item>
+                    <div className="legend">
+                      <span className="urgency-color"></span>{" "}
+                      <span className="legend-text">Urgent</span>
+                    </div>
+                  </Grid>
+                  <Grid item>
+                    <div className="legend">
+                      <span className="normal-color"></span>{" "}
+                      <span className="legend-text">Normal</span>
+                    </div>
+                  </Grid>
+                  <Grid item>
+                    <div className="legend">
+                      <span className="medium-color"></span>{" "}
+                      <span className="legend-text">Medium</span>
+                    </div>
+                  </Grid>
                 </Grid>
-                <Grid item>
-                  <div className="legend">
-                    <span className="normal-color"></span>{" "}
-                    <span className="legend-text">Normal</span>
-                  </div>
-                </Grid>
-                <Grid item>
-                  <div className="legend">
-                    <span className="medium-color"></span>{" "}
-                    <span className="legend-text">Medium</span>
-                  </div>
-                </Grid>
-              </Grid>
-            </div>
+              </div>
 
-            <style>
-              {`
+              <style>
+                {`
                 .urgency {
                   background-color:#FFA07A; /* Red */
                 }
@@ -148,83 +157,39 @@ export default function SpecialMsgDisplay() {
                 }
 
               `}
-            </style>
-            <DataGrid
-              autoHeight
-              rows={customDataset}
-              columns={[
-                ...VISIBLE_FIELDS.map((field) => ({
-                  field,
-                  headerName: field,
-                  width: 250, // Adjust width as needed
-                })),
-                {
-                  field: "delete",
-                  headerName: "Delete",
-                  width: 100,
-                  renderCell: (params) => (
-                    <Button
-                      variant="outlined"
-                      color="error"
-                      onClick={() => handleDelete(params.row.id)}
-                    >
-                      Delete
-                    </Button>
-                  ),
-                },
-              ]}
-              components={{ Toolbar: GridToolbar }}
-              getRowClassName={getRowClassName}
-            />
+              </style>
+              <DataGrid
+                autoHeight
+                rows={customDataset}
+                columns={[
+                  ...VISIBLE_FIELDS.map((field) => ({
+                    field,
+                    headerName: field,
+                    width: 250, // Adjust width as needed
+                  })),
+                  {
+                    field: "delete",
+                    headerName: "Delete",
+                    width: 100,
+                    renderCell: (params) => (
+                      <Button
+                        variant="outlined"
+                        color="error"
+                        onClick={() => handleDelete(params.row.id)}
+                      >
+                        Delete
+                      </Button>
+                    ),
+                  },
+                ]}
+                components={{ Toolbar: GridToolbar }}
+                getRowClassName={getRowClassName}
+              />
+            </div>
           </div>
         </div>
-        <div style={{ marginTop: "20px", color: "black" }}>
-          {/* Add phone number and email here */}
-          <Typography variant="h8" style={{ color: "black" }}>
-            Emergency Contact - Midwife
-          </Typography>
-          <Grid container spacing={2} alignItems="center">
-            <Grid item>
-              <IconButton
-                component={Link}
-                href="tel:+94767087009"
-                style={{ color: "#2A777C" }}
-              >
-                <PhoneIcon />
-              </IconButton>
-            </Grid>
-            <Grid item>
-              <Typography variant="body1">
-                <Link href="tel:+94767087009" style={{ color: "#2A777C" }}>
-                  +94767087009
-                </Link>
-              </Typography>
-            </Grid>
-          </Grid>
-          <Grid container spacing={2} alignItems="center">
-            <Grid item>
-              <IconButton
-                component={Link}
-                href="mailto:nimanthijayathilaka99@gmail.com"
-                style={{ color: "#2A777C" }}
-              >
-                <EmailIcon />
-              </IconButton>
-            </Grid>
-            <Grid item>
-              <Typography variant="body1">
-                <Link
-                  href="mailto:nimanthijayathilaka99@gmail.com"
-                  style={{ color: "#2A777C" }}
-                >
-                  nimanthijayathilaka99@gmail.com
-                </Link>
-              </Typography>
-            </Grid>
-          </Grid>
-        </div>
-        ;
-      </div>
-    </Box>
+      </Box>
+      {showSuccessAlert && <SuccessAlert setAlert={setShowSuccessAlert} />}
+    </>
   );
 }
